requirements.txt
browser-use==0.7.0
python-dotenv>=1.0.1
